Bootstrap Consul raft with Containerbuddy
Simplify bootstrap by using Containerbuddy health checks to join a
cluster that's been primed with `-bootstrap-expect`.

Our new start script starts one instance w/ BOOTSTRAP_HOST unset and
then uses that host IP for BOOTSTRAP_HOST on subsequent scale-ups.

When an instance does its health check, it checks to see if it's
already a member of a raft. If not, it checks if it has the env var
BOOTSTRAP_HOST set. If not, it starts the raft by joining itself.
If so, it joins the BOOTSTRAP_HOST.