Implement raft-test.js test of raft status
We query Consul to determine the status of the raft by getting a list of
peers and the leader. We compare these values against the expected list.
If they match exactly and one of those nodes is the leader, the test passes.
If not, we retry twice and then return error to the callback if the raft
still has no healed.