Use CNS to bootstrap the Consul cluster
A node's health check will check to see if it is a member of a raft.
If it has not peers then it will try to join the raft via the CNS
svc record. Once a node is connected to at least one other peer it'll
get the rest of the raft via the Consul LAN gossip.